// packages/azdo-tasks/teamsfx-cli-task/buildandreleasetask/src/enums/singleOptions.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT license.

// eslint-disable-next-line no-shadow
export enum SingleOptions {
  Verbose = 'verbose',
  Version = 'version',
  Debug = 'debug',
  Help = 'help',
  Action = 'action',
  Tenant = 'tenant',
  Folder = 'folder',
  Subscription = 'subscription',
  Interactive = 'interactive',
  Scratch = 'scratch',
  HostType = 'host-type',
  SpfxFrameworkType = 'spfx-framework-type',
  SpfxWebpartName = 'spfx-webpart-name',
  SpfxWebpartDesp = 'spfx-webpart-desp',
  WayToRegisterBot = 'way-to-register-bot',
  BotId = 'bot-id',
  BotPassword = 'bot-password',
  ProgrammingLanguage = 'programming-language',
  AppName = 'app-name',
  Samples = 'samples',
  SqlAdminName = 'sql-admin-name',
  SqlPassword = 'sql-password',
  OpenApiDocument = 'open-api-document',
  ApiPrefix = 'api-prefix',
  ApiVersion = 'api-version',
  Environment = 'environment',
  Endpoint = 'endpoint',
  RootPath = 'root-path'
}